Django_app/requirements.txt
alembic==1.13.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
python-dotenv>=0.19.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
requests>=2.26.0
fhirclient>=4.2.1
twilio>=7.0.0
openai>=0.27.0
dateparser
django>=3.2,<4.0
djangorestframework>=3.12.0
redis>=4.0.0
celery>=5.2.0
aioredis>=2.0.0
python-dateutil>=2.8.2
pytz>=2021.3
uvicorn>=0.15.0
starlette>=0.27.0
presidio-analyzer==2.2.357
langdetect==1.0.9
sentence-transformers==2.2.2
huggingface-hub==0.20.3
transformers==4.30.0